test: migrate rh_ip unit tests from TestCase to pytest
Move tests/unit/modules/test_rh_ip.py to
tests/pytests/unit/modules/test_rh_ip.py, converting the legacy
unittest.TestCase style (LoaderModuleMockMixin, self.assertX) to modern
pytest (configure_loader_modules fixture, plain asserts, pytest.raises).
Faithful style-only migration: all 26 tests preserved and passing; the
legacy file (which no longer runs under the current pytest/py3.12 test
env) is removed.

This lets the rh_ip test suite run in the maintained pytests tree, ahead
of adding NetworkManager-aware provider selection for #54791.
